+++ CHANGELOG.md
@@ -30,6 +30,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - Added support for session referer debug parameter in CampaignDebugger, improved output and usability. remp/remp#1459
 - Added support for session debug parameter in CampaignDebugger, improved output. remp/remp#1459
 - Changed banner close button clickable targets according to WCAG recommendation. remp/remp#1458
+- Added snippet usage information to the snippet edit page, showing which banners and other snippets reference the snippet (including direct vs transitive usage). remp/remp#1453
 
 ### [Mailer]